move CLI command files into cli/commands/ subdirectory
Move all 26 command registration files from assistant/src/cli/ into
assistant/src/cli/commands/ so that cli/ contains only utilities,
infrastructure, and entry points (program.ts, utils.ts, reference.ts,
ipc-client.ts, email-guardrails.ts, main-screen.tsx).

Update all relative imports in moved files (../X → ../../X, ./utils →
../utils), program.ts imports (./X → ./commands/X), and 6 test files
that imported command modules directly.
