+++ assistant/src/cli/commands/default-action.ts
@@ -1,8 +1,8 @@
 import type { Command } from "commander";
 
-import { startCli } from "../cli.js";
-import { shouldAutoStartDaemon } from "../daemon/connection-policy.js";
-import { ensureDaemonRunning } from "../daemon/lifecycle.js";
+import { startCli } from "../../cli.js";
+import { shouldAutoStartDaemon } from "../../daemon/connection-policy.js";
+import { ensureDaemonRunning } from "../../daemon/lifecycle.js";
 
 export function registerDefaultAction(program: Command): void {
   program.action(async () => {
