+++ assistant/src/cli/commands/dev.ts
@@ -3,8 +3,8 @@ import { join } from "node:path";
 
 import type { Command } from "commander";
 
-import { getDaemonStatus, stopDaemon } from "../daemon/lifecycle.js";
-import { getCliLogger } from "../util/logger.js";
+import { getDaemonStatus, stopDaemon } from "../../daemon/lifecycle.js";
+import { getCliLogger } from "../../util/logger.js";
 
 const log = getCliLogger("cli");
