+++ assistant/src/cli/program.ts
@@ -3,32 +3,32 @@ import { createRequire } from "node:module";
 import { Command } from "commander";
 
 import { registerHooksCommand } from "../hooks/cli.js";
-import { registerAmazonCommand } from "./amazon.js";
-import { registerAuditCommand } from "./audit.js";
-import { registerAutonomyCommand } from "./autonomy.js";
-import { registerChannelVerificationSessionsCommand } from "./channel-verification-sessions.js";
-import { registerChannelsCommand } from "./channels.js";
-import { registerCompletionsCommand } from "./completions.js";
-import { registerConfigCommand } from "./config.js";
-import { registerContactsCommand } from "./contacts.js";
-import { registerCredentialsCommand } from "./credentials.js";
-import { registerDefaultAction } from "./default-action.js";
-import { registerDevCommand } from "./dev.js";
-import { registerDoctorCommand } from "./doctor.js";
-import { registerEmailCommand } from "./email.js";
-import { registerInfluencerCommand } from "./influencer.js";
-import { registerKeysCommand } from "./keys.js";
-import { registerMapCommand } from "./map.js";
-import { registerMcpCommand } from "./mcp.js";
-import { registerMemoryCommand } from "./memory.js";
-import { registerNotificationsCommand } from "./notifications.js";
-import { registerOAuthCommand } from "./oauth.js";
-import { registerPlatformCommand } from "./platform.js";
-import { registerSequenceCommand } from "./sequence.js";
-import { registerSessionsCommand } from "./sessions.js";
-import { registerSkillsCommand } from "./skills.js";
-import { registerTrustCommand } from "./trust.js";
-import { registerTwitterCommand } from "./twitter.js";
+import { registerAmazonCommand } from "./commands/amazon.js";
+import { registerAuditCommand } from "./commands/audit.js";
+import { registerAutonomyCommand } from "./commands/autonomy.js";
+import { registerChannelVerificationSessionsCommand } from "./commands/channel-verification-sessions.js";
+import { registerChannelsCommand } from "./commands/channels.js";
+import { registerCompletionsCommand } from "./commands/completions.js";
+import { registerConfigCommand } from "./commands/config.js";
+import { registerContactsCommand } from "./commands/contacts.js";
+import { registerCredentialsCommand } from "./commands/credentials.js";
+import { registerDefaultAction } from "./commands/default-action.js";
+import { registerDevCommand } from "./commands/dev.js";
+import { registerDoctorCommand } from "./commands/doctor.js";
+import { registerEmailCommand } from "./commands/email.js";
+import { registerInfluencerCommand } from "./commands/influencer.js";
+import { registerKeysCommand } from "./commands/keys.js";
+import { registerMapCommand } from "./commands/map.js";
+import { registerMcpCommand } from "./commands/mcp.js";
+import { registerMemoryCommand } from "./commands/memory.js";
+import { registerNotificationsCommand } from "./commands/notifications.js";
+import { registerOAuthCommand } from "./commands/oauth.js";
+import { registerPlatformCommand } from "./commands/platform.js";
+import { registerSequenceCommand } from "./commands/sequence.js";
+import { registerSessionsCommand } from "./commands/sessions.js";
+import { registerSkillsCommand } from "./commands/skills.js";
+import { registerTrustCommand } from "./commands/trust.js";
+import { registerTwitterCommand } from "./commands/twitter.js";
 
 const require = createRequire(import.meta.url);
 const { version } = require("../../package.json") as { version: string };
